Fix sidebar to be collapsed in mobile view
Motivation:

The sidebar is hidden by default in mobile view, and it is able to
toggle the visibility.
The problem is that it is not hidden again when a user clicks a link
in the sidebar. A user has to close it all the time, and it is definitely
bad user experience.

Modification:

Collapse the sidebar when clicking a link in the sidebar in mobile view

Result:

Better user experience for mobile doc service