Implement the public methods in Object for dynamic proxy
Motivation:

The InvocationHandler implementation in ClientBuilder does not handle
the public methods in Object class: toString(), hashCode() and equals()

Whenever a user or a debugger attempts to call one of those methods, the
caller will observe an unexpected behavior from the InvocationHandler.

Modifications:

- Extract the InvocationHandler implementation to a dedicated class:
  - ClientInvocationHandler
- Implement the public methods in Object
- Miscellaneous
  - Fix NPE in SimpleHttpRequest.toString() when content is null, which
    can happen when a debugger tries to call toString() before
    construction is finished
  - Fix NPE when the proxied method has no argument

Result:

Application does not fall apart even when it calls a public method
declared in java.lang.Object.