Redirect from /path to /path/ instead of 404
Related: #51

Motivation:

Most web servers send a redirect when a client sends a non-existent path
that doesn't end with '/' and the path + '/' exists. We should do the
same in our HttpServerHandler

Modifications:

- Add handleNonExistentMapping() method to handle the case where
  findServiceConfig(path) returns an empty mapping properly
- Add redirect() to send a redirect response
- Split channelRead() into smaller methods for readability
  - handleHttp2Settings()
  - handleRequest()
    - handleNonExistentMapping()
    - handleInvocationPromise()
    - handleDecodeFailure()
- Reordered methods so that related methods are grouped together

Result:

You don't need to be careful anymore when you enter a URL that ends with
'/'. e.g. /docs