More robust resolution/rejection of a promise
Motivation:

When a client does not receive a response or a server fails to process a
request within time limit, an invocation promise is rejected regardless
the final outcome of the invocation.

In such a case, a ServiceInvocationHandler or a remote invoker
implementation might attempt to resolve or reject the promise which is
already rejected with TimeoutException. Handling this sort of situations
is very common, and thus we should provide it as part of Armeria.

Modifications:

- Add ServiceInvocationContext.resolvePromise() and rejectPromise() to
  provide a simple and robust way to resolve or reject a promise in a
  ServiceInvocationHandler
- Use resolvePromise() and rejectPromise() wherever possible.
- Miscellaneous:
  - Fix a bug where WriteTimeoutException is raised rather than
    ResponseTimeoutException
  - Fix a bug where HttpSessionHandler.failPendingResponses() does not
    fail anything at all
  - Make TimeoutException traceless to reduce its instantiation overhead

Result:

- Robustness
  - Fixes potential issues when attempting to resolve a rejected promise.
  - Late responses are now silently discarded to avoid log message
    flooding