Add a simpler way to decorate a client / Enrich ClientBuilder
Motivation:

When a user attempts to decorate a client, it is much more verbose than
decorating a service. A user should be given with a simpler way to
configure a client when using ClientBuilder.

Modifications:

- Add a new interface 'Client' which is a client-side counterpart of
  'Service'
  - Add SimpleClient and DecoratingClient
- Add various setter methods incloding decorator() to ClientBuilder
- Add LoggingClient and hide LoggingClientCodec
- Remove the ability to decorate an InvocationHandler, because it
  seems that RemoteInvoker decorators can do the exact same job.
- Miscellaneous:
  - Make LoggingClientCodec extend DecoratingClientCodec
  - Make AbstractOptions and its subtypes accept
    Iterable<AbstractOption> as well

Result:

A user can build a client in a more fluent and less verbose way:

  builder.decorator(LoggingClient::new)
         .decorator(MyDecorator::new)
         .build();