Send the correct 'Host' header when upgrading.
Related: #98

Motivation:

HttpSessionChannelFactory keeps the cached map from SessionProtocol to
Bootstrap. The cached Bootstrap configures its new Channels to send the
name of the target host with an H1C-to-H2C upgrade request.

Unfortunately, The cached Bootstrap does not determine the target host
name dynamically but just uses the initially specified one, and this
makes a client send a wrong 'Host' header when a pool creates a new
connection with a different target host name and with the same
SessionProtocol.

Modifications:

- HttpConfigurator doesn't require a 'host' parameter anymore. It
  retrieves the target host name by intercepting the 'connect()'
  operation.
- Add HttpHostHeaderUtil that creates 'Host' header value correctly
  - Use HttpHostHeaderUtil.hostHeader() in HttpConfigurator and
    HttpSessionHandler to reduce code duplication
- Miscellaneous:
  - Replace HttpRemoteInvoker.poolKey() with direct constructor call
  - Fix the API documentation of ServiceInvocationContext.host()
  - Reorganize the methods and inner classes in HttpConfigurator
  - Remove ByteBufOverHttpCodec which is not used anymore

Result:

- Armeria client sends the correct 'Host' header when it sends a upgrade
  request.
- Description of ServiceInvocationContext.host() is not misleading
  anymore.