Reduce the unnecessarily noisy/long exceptions
Motivation:

When Armeria is under heavy load that it cannot keep up with, exceptions
will be triggered, and some of them do not really need logging.

Here's the list of the exceptions which are triggered very usually in
case of socket disconnection due to timeout / connection reset by peer
errors:

- IOException: connection reset by peer / broken pipe
- Http2Exception$StreamException: Stream closed
- ClosedChannelException
- ClosedSessionException

Also, an attempt to close a HTTP/2 connection with pending writes can
cause unnecessarily long stack trace because of the following steps they
involve:

1. Channel.close() triggers AbstractHttpToHttp2ConnectionHandler.close().
2. AbstractHttpToHttp2ConnectionHandler.close() triggers Http2Stream.close().
3. Http2Stream.close() fails the promise of its pending writes.
4. The failed promise notifies HttpServerHandler.CLOSE_ON_FAILURE.
5. HttpServerHandler.CLOSE_ON_FAILURE calls Channel.close().
6. Repeat from the step 1.

More the pending writes, longer the stack trace.

Modifications:

- Add Exceptions, a utility class that filters out the exceptions
  mentioned above
  - Use Exceptions in ServiceInvocationContext and ChannelHandler
    implementations to reduce the amount of noisy log messages
- Do not call Channel.close() if HttpToHttp2Connection.close() has been
  invoked already or the Channel has been closed already

Result:

- Much less noise when Armeria server is overloaded
- The bad call cycle that was cause unnecessarily long stack trace is gone.