Make EventLoop available to RemoteInvoker.invoke()
Motivation:

EventLoop to perform the actual invocation is determined at the
HttpRemoteInvoker.invoke() method, which is the last moment right before
the invocation.

It means decorating RemoteInvokers does not have an access to the
EventLoop, nevertheless the EventLoop can be determined from the
beginning of the invocation.

A decorating RemoteInvoker with an access to an EventLoop could create
Future/Promise and schedule a task, which will be useful when building
an advanced decorator such as a circuit breaker.

Modifications:

- Add EventLoop parameter to RemoteInvoker.invoke()
- Determine the EventLoop to perform the invocation in
  ClientInvocationHandler instead of in HttpRemoteInvoker
  - Move HttpRemoteInvoker.eventLoop() to ClientInvocationHandler
- Reorder the methods in HttpRemoteInvoker for readability

Result:

A decorating RemoteInvoker now has an access to the EventLoop which will
perform the actual invocation.