Use the HTTP/2 preface string for H1C-to-H2C upgrade
Related: #142

Motivation:

Armeria client currently sends a 'HEAD / HTTP/1.1' request to issue a
protocol upgrade, as recommended by the RFC, but it can cause heavy load
on the server side if the service bound at / does not implement HEAD
method optimally. When many Armeria clients creates a lot of connections
concurrently, the server might not respond to the upgrade request in a
timely manner.

Modifications:

- Add RemoteInvokerOption.USE_HTTP2_PREFACE which is true by default
  - When set to true, Armeria client uses 'PRI * HTTP/2.0' preface
    string instead of the old 'HEAD / HTTP/1.1' request to upgrade from
    H1C to H2C.
  - A user can specify a system property to change the default value of
    the USE_HTTP2_PREFACE option.
- Implement the H1C-to-H2C upgrade that uses the HTTP/2 preface string,
  while retaining the old behavior when USE_HTTP2_PREFACE is set to
  false

Result:

- Fixes #142
- Armeria client does not produce load on the server accidentally.
- HTTP/2 clients have more upgrade options.