Fix backward compatibility with older Ameria versions
Motivation:

1)

The old Armeria server does not handle an HTTP request whose path is '*'
correctly. As a result, sending an HTTP/2 connection preface string will
make the server close the connection immediately.

HttpConfigurator.DowngradeHandler does not handle the case where the
connection is closed before the first 4 bytes are received, and thus it
fails to downgrade to H1C.

2)

Armeria server raises an unexpected exception when the path of an HTTP
request is not an absolute path rather than responding with 400 Bad
Request.

Also, it doesn't handle 'OPTIONS * HTTP/1.1' requests because of the
same reason.

Modifications:

- Make sure DowngradeHandler switches to H1C when less than 4 bytes were
  received before disconnection.
- Use 'HEAD / HTTP/1.1' again as the default upgrade mechanism for
  backward compatibility
- Respond with a '400 Bad Request' response when a request does not have
  an absolute path
- Handle 'OPTIONS * HTTP/1.1' specially
- Miscellaneous:
  - Move HttpSessionChannelFactory.HEALTH_CHECKER to HttpRemoteInvoker
    because it's used only there
  - Add documentation for HttpSessionHandler fields

Result:

- Armeria client works fine with the old Armeria servers again.
- No more IllegalArgumentException on non-absolute path requests
- Armeria server handles 'OPTIONS *' correctly