Fix NoSuchMethodError when using Tomcat 7
Motivation:

A user was able to use TomcatService with Tomcat 7 by instantiating a
TomcatService via the forConnector() factory method.

However, it does not work anymore because e06096dd115d4a7ebf7f28f6315f2ddec05c50b2
introduced the call to the method that does not exist in Tomcat 7 but
only in Tomcat 8+.

Modifications:

- Do not call Server.getCatalinaBase() if the current Tomcat version is
  less than 8.
- Do not attempt to start Server when its state is STARTED, STARTING,
  STARTING_PREP.

Result:

A user can create a TomcatService with a Tomcat 7 Connector again.