Do not send a GET request with the transfer-encoding header
Motivation:

When a user sends a GET/HEAD/DELETE/TRACE request which is supposed to
have empty content, Armeria sets the 'transfer-encoding: chunked' header
followed by a zero-length chunk.

It is not strictly prohibited by HTTP/1.1 to send such a request, it is
not desirable to do so and some servers do not accept it.

Modifications:

- Add RichPublisher.isEmpty() which tells if it is sure that the stream
  did not and will not publish anything at all.
  - Set neither 'transfer-encoding' or 'content-length' if
    HttpRequest.isEmpty() returns true
- Add HttpRequest.of(HttpHeaders) that create an empty-content request
- Add a test case
- Miscellaneous:
  - Fix PublisherWithCloseFuture.abort()

Result:

Armeria now sends more sensible GET/HEAD/DELETE/TRACE requests.