Support client-side TMultiplexedProtocol
Related: #242 and #240

Motivation:

Community request for TMultiplexedProtocol. TMultiplexedProtocol is a
simple variation of Thrift which prepends a service name to the method
name field (e.g. serviceName:methodName) so that a single endpoint can
serve the requests for different Thrift services

Modifications:

- Add ClientRequestContext.fragments()
- Use the fragment part of the request URI as the service name of a
  multiplexed Thrift call
- Add THttpClient.executeMultiplexed()

Result:

A user can make a multiplexed Thrift call:

    // The service name is 'foo'.
    FooService.Iface client =
            Clients.newClient("tbinary+http://127.0.0.1:8080/thrift#foo",
                              FooService.Iface.class);