Add 'sticky HTTP headers' checkbox in the DocService debug form
Motivation:

When sending a request via the DocService debug form, it's often the
case different requests have identical HTTP headers (e.g. authentication
token.)

It will be useful if there's a way to keep the HTTP headers in the debug
form when a user switches to other functions, so that he or she does not
need to enter the authentication token again and again.

Modifications:

- Add 'Use these HTTP headers for all functions.' checkbox, which makes
  the HTTP headers in the debug form is retained when switching to other
  function.
- Add 'http_headers_sticky' query parameter which preserves the checkbox
  state

Result:

- Fixes #276
- @taicki is happy. :-)