Do not pop the current RequestContext too early when a Service throws an unexpected exception
Motivation:

When a Service throws an unexpected exception in its `service()` method,
HttpServerHandler.handleRequest() will log the exception at WARN level.

When the exception is logged, the thread-local RequestContext must be
set so that RequestContextExportingAppender picks it up.

However, we currently pop the RequestContext too early, and thus
RequestContextExportingAppender does not see the thread local
RequestContext.

Modifications

- Pop the current RequestContext as late as possible
- Fix potential race condition in RequestContextExportingAppender

Result:

- An unexpected exception from a Service will be logged with more
  information.
- RequestContextExportingAppender will not cause other Appender to throw
  ConcurrentModificationException by making a copy