Add structured request/response logging support with integration to Kafka (#324)
Motivation:

Typically, when we build an RPC based application, the access log
doesn't tell many things. We've found it is very useful to introduce
service log (detailed structured log of each request and response),
expecting various ways to utilize it:

- Send RPC logs to stats environment and analize by various ways such
  as batch processing or stream processing to get insights of the
  actual usage of the service.
- Troubleshooting by inspecting collected logs. It's much easy to do as
  everything of the "actual request and response" are contained in each
  entry.
- Build a "shadow" environment of production cluster and replay the
  actual production traffic for performance testing, troubleshooting or
  etc.

Modifications:

- Add `ServiceLogLoggingService` which decorates a service with
  enabling fully detailed request/response logging in structured format.
- Add some classes as a framework to let users implement service log
  logging in the unified way.
- Add a new module `armeria-kafka` which implements a backend of
  service log logging with integration to Kafka.

Design:

There are a few points that are considred to be customizable by users
in order to support various different environment and contexts.

*Logging backend*

There are number of middlwares which are invented to collect a logs
from application instances that are located in different servers.
Apache Kafka, Logstash, Fluentd are some.

As the first step this PR will add a support for Apache Kafka
integration. However, the interface which represents logging backend
stay underlying middleware agnostic so users can write their own output
by their own if necessary, and hopefully they might be contribute it as
a module of Armeria.

*Log message format*

Different organization uses differnet serialization formats. Since the
purpose of this PR is to introduce structured content logging, we're
focusing on serialization format of structured data, but even with
that, we can list many of them. JSON, Protocol buffers and Thrift are
some.

JSON is an easy choice but some organizations(including us) may want
to use much schema oriented serialization format such as Protocol
Buffers and Thrift in order to achieve uniformed wire format across the
organization, and/or much higher SerDe level capability such as
constraints and schema evolution.

We may support JSON as the default in order to let users try this
feature with ease, but the interface should stay open for any
serializer plugin.

In fact, at this moment, I decided to let each backends to supply their
own specification for serializer interface. As different backend wants
different result types(byte[]? String?), I think the desired result
type of serializeation should stick with each backends.

*Contents of a log entry*

We can predefine a information preset of each protocol-oriented request
and response. For instance, in case of Thrift RPC based application, we
may wanna log the following information in general:

- Thrift service name
- Thrift method name
- Thrift arguments
- Thrift response or exception
- The timestamp of when the request received
- The duration taken to process the request

Those information are able to extract and construct from `RequestContext`
and `ResponseLog` which are given to the decorating service, but maybe
they aren't all that users want to log.

Therefore we should provide a way to customize the content of a log
entry. The expected type of entry also differs among different protocols,
so the log entry builder should also be a parameter of service log logging.

Working example:

Here's an example of a log output when I called ArmeriaService.hello:

```
./kafka-console-consumer.sh ... --topic armeria-service-log-test
{
  "timestampMs": 1478243846990,
  "responseTimeMs": 35,
  "thriftServiceName": "com.linecorp.armeria.service.core.thrift.v1.ArmeriaService",
  "thriftMethodName": "hello",
  "thriftCall": {
    "header": {
      "name": "hello",
      "type": 1,
      "seqid": 0
    },
    "args": {
      "name" : "kawamuray"
    }
  },
  "thriftReply": {
    "header": {
      "name": "hello",
      "type": 2,
      "seqid": 0
    },
    "result": {
      "success" : "Hello kawamuray"
    },
    "exception":null
  }
}
```