Revamp RequestLog API (#343)
* Revamp RequestLog API

Motivation:

RequestLog API currently notifies the log consumer only when a Request
or a Response is fully processed. This does not work well with a
streaming request that can take potentially long time when the consumer
wants to get certain information sooner.

A good example would be the implementation of 'activeRequest' counter.
DropwizardMetricConsumer increments the 'activeRequest' counter in
onRequest() handler method and decrements in onResponse(). This is OK
for Thrift calls where request processing takes small amount of time,
but not OK for other streaming calls such as file upload.

There should be a way to let a log consumer tell the API what properties
the log consumer is interested in and RequestLog should notify the
consumer as soon as the interested properties are available.

Modifications:

- Rewrite RequestLog and RequestLogBuilder
  - ResponseLog and ResponseLogBuilder are merged into RequestLog and
    RequestLogBuilder for simplicity
  - Add RequestLogAvailability to define the availability of properties
- Remove LogCollecting{Client,Service} because it doesn't do anything
  useful anymore with the new API
- Rename AbstractRpc{Request,Response} to DefaultRpc{Request,Response}
- Remove ThriftCall and ThriftReply and use DefaultRpc{Request,Response}
  - There was no reason to introduce these classes from the beginning...
- Rename ApacheThrift{Call,Reply} to Thrift{Call,Reply}
- Add the documentation about the new API
  - Update CSS a little bit to make things look better
  - Add a placeholder page for @kawamuray's Kafka structured logging
  - Do not use Java classes in conf.py so that the site can be generated
    with CPython
- Miscellaneous:
  - Make the formatter methods in TextFormatter append the raw value
    when the value is formatted into an approximated (or human-friendly)
    form

Result:

- Possible to get the necessary information sooner
- Cannot add an attribute to RequestLog anymore
  - May be both pro and con; will revisit if there's demand
- RequestLog API now has documentation.

* Follow-up commit

- Do not set requestContent to null implicitly on serializationFormat(NONE)
- Make AbstractHttpService set requestContent and serializationFormat
  automatically
- Add RequestLogBuilder.is{Request,Response}ContentDeferred()
- Update documentation