Have shadow plugin relocate source references too by defining a separate task (#395)
Have shadow plugin relocate source references too by defining a separate task for it.
Only run shading and proguarding when building a JAR.

When doing some recent work, I found it annoying that

- I had no sources for relocated dependencies I wanted to use and understand
- Running tests from command line always ran a very slow proguard task

The main reason for this seemed to be us only running shadow on dependencies, not on the armeria source itself. This is probably because we didn't want shadow to bundle all the dependencies as it usually does. Indeed, I couldn't find a good way to only bundle the relocated deps along with proguard - it was either all or nothing.

I found this approach that adds an additional shadow step that runs on the source with no bundling and overwrites the jar contents with it. It appears to be working. I have done the rewriting in all packages since currently it seems we have a style of having internal dependencies in core, and having other packages like thrift use them. If we ever had subpackage-specific internal dependencies, there'd need to be similar tasks added to the subpackage but otherwise should work ok.

Result:

Easier development since weirdness is all moved to the packaging step.

Extra change:

I moved the trimmed internal dependencies into the core build directory.