Allow the lazy instantiation of Jetty SessionIdManager
Motivation:

Since Jetty 9.4, DefaultSessionIdManager requires a Jetty Server
instance as a constructor parameter, which means a user cannot create it
until a Server instance is actually created. It is currently impossible
to achieve this with JettyService because JettyServiceBuilder requires
a SessionIdManager instance rather than its factory.

Modifications:

- Add JettyServiceBuilder.sessionIdManagerFactory() so that a user can
  create a SessionIdManager lazily
- Miscellaneous:
  - Use toStringHelper instead of manual concatenation

Result:

Better customizability of Jetty Server