Clean up test code and its dependency relationship
Motivation:

We relied on a hack that makes the 'java-shared' or 'resources-shared'
directories are compiled into multiple output directories to share the
test code and test resources.

Although it worked without a problem when built from CLI, it confused
IDEs such as IntelliJ IDEA because it seems to assume one source
directory is mapped into one output directory.

Modifications:

- Add a new sub-project 'testing'
- Move all test-related classes and resources there, notably:
  - core/src/test/java-shared
  - tomcat/src/test/*-shared
  - settings/logback
- Make all sub-projects depend on 'testing'
- Improve the overall experience of Server testing
  - Replace AbstractServerTest with ServerRule
  - Add SelfSignedCertificateRule
  - WebAppContainerTest now makes use of ServerRule
  - Add WebAppContainerTest.webAppRoot() so that we can find the docBase
    of the test webapp easily
- Fix a bug in tomcat8.0/build.gradle where it uses 'sourceSets',
  confuses IntelliJ IDEA and resulting in build errors.

Result:

- Fixes #427
- Better code organization all in all