Add support for grpc-web draft spec. (#522)
Browsers cannot use normal grpc because they don't support trailers, so a grpc-web spec exists that is basically the same as grpc but encodes trailers into the response body. It also talks about decoupling the protocol from HTTP/2, but armeria already does that :D

Spec:
https://github.com/grpc/grpc/blob/master/doc/PROTOCOL-WEB.md

No reference implementation, but closest is at:
https://github.com/improbable-eng/grpc-web

Verified interop with the example in that repo.