Add support for subscribing to StreamMessage with or without pooled objects (#528)
Currently, it is possible for decorators to subscribe to a response and return a different one without releasing pooled objects. This makes the default behavior of subscription do a copy from pooled to unpooled to protect unknowing decorators from memory leaks. As modifying the response content in a decorator is quite rare, this copy should rarely happen in practice.

Also adds a simple benchmark of pooled vs unpooled of medium size buffers.