Rearrange callbacks for server error case to match GRPC semantics. (#541)
While examining error handling, I realized that GRPC server does not actually send an error response to the client in cases of transport errors (request too large / invalid, deadline) - it just resets the stream with HTTP/2 CANCEL. Note, this doesn't mean I don't need #521 as I would still like to be able to notify the server's business logic of cancellation and clear up stream resources.

This PR makes sure that non-business-logic exceptions are not returned to the client, but just result in a reset. It also renames to transportStatus and closeListener to match the naming of upstream grpc-java to make comparing the implementations simpler. 

Also:
- Use same translation of netty error codes to status as grpc
- Remove unused method
- Adds DisableOnDebug to test timeout. My previous attempt at this, if you remember it, failed since it doesn't affect `Test(timeout = ` pattern, but for this test I'm only using a global timeout. This is probably a good pattern to use generally in armeria since it makes debugging much easier and removes some of the noise on timeouts.