Inject a query string parameter as an annotated method argument
Motivation:

The other framework, such as Spring framework, provides query string parameter injection basically.
So it is better to support the feature on Armeria so that a user writes less codes.
We are going to support parameter injection for a query string of HTTP GET requests and url-encoded form data of HTTP POST requests.

Modifications:
- Add @Param annotation for parameter injection.
- Add HttpParameters interface to inject every parameters.
- Replace @PathParam with @Param annotation. @PathParam is removed.
- Respond "400 Bad Request" if the request has not suffcient arguments or has an invalid argument. (was: 500 Internal Server Error)
- Allow path pattern as a value of HTTP method annotations. e.g, @Get("/path")
- Add @Optional annotation to specify an optional parameter and its default value. e.g, @Param("name") @Optional("thomas") String name

Result:
- A user is able to get parameter easily on his/her annotated service methods.