Support path parameter names that start with a digit (#587)
Motivation:

For example, the following will fail:

    PathMapping.of("/{0}");

.. because we use the path parameter name a user specified in verbatim
when building a regular expression. We don't really need to use named
capturing groups in the generated regular expression. We can simply keep
an array which contains the path parameter names in the order of
appearance and then map the capturing group number into a path parameter
name at O(1) time.

Modifications:

- Use numbered capturing groups for the regular expression generated by
  DefaultPathMapping

Result:

- Fixes #586