Add StreamMessageDuplicator
Motivation:

If a stream message is, once published to a subscriber, next subscriber cannot subscribe
the data again. A user may want to see just the head of data and decide to request other
subscriber to do more action according to the data of the head.
In that case, another subscriber cannot see the data consumed already.
We needed to provide the way that a publisher can be subscribed by multiple
subscribers, so a user can use the elements in different places.

Modifications:

- Add StreamMessageDuplicator that wraps a StreamMessage and spawns
  Children stream messages can be subscribed one by one
- Add HttpResponseDuplicator and HttpRequestDuplicator

Result:

- A user can subscribe a publisher with multiple subscribers.

To-do:

- Limit the size of the buffered data in the factory