Restore prefix mapping stripping to allow it to work together with aribtrarily mapped PathMappingResult. (#593)
Currently, the way to get a stripped prefix path is `pathWithoutPrefix`, but we support internal redirects / arbitrary mapping with `PathMappingResult.path`, which is returned by `path`. This makes it mostly impossible to have a service respect both types of mapping.

This PR restores the behavior of prefix mapping stripping the input and allowing services to use the `mappedPath` which has been undeprecated.

We're basically back to before the HTTP-only PR, with `mappedPath` returning `PathMappingResult.path`.

Also removes `pathWithoutPrefix` since it's not needed with this change, and is very short lived so probably not worth deprecating.