Ensure an HTTP object is requested after reading headers. (#608)
I had misunderstood that the subscription demand is only needed between data frames, but on the client-side it's also needed after the initial headers. Without correctly requesting a frame, the stream hangs.

Currently, if onSubscribe happens to be run before request, than two frames are requested at subscription time which allows this error to be hidden, but this is not guaranteed order, and in the opposite order, only one frame is correctly requested (for some reason I reliably trigger the latter behavior on a server I am developing, not sure yet why I'd see such different behavior than the unit tests). Anyways, I have tweaked the initial request semantics to not allow this dual-behavior for clarity.