Make the default blockingTaskExecutor utilize more than one thread (#620)
Motivation:

The default blockingTaskExecutor uses 0 core threads and an unbounded
task queue. This will make blockingTaskExecutor single-threaded because
ThreadPoolExecutor will spawn a new thread only when a task queue
rejects a task, which will never happen with an unbounded task queue.

Modifications:

- Set the number of the core threads to the maximum and allow the core
  threads to be timed out instead.

Result:

The default blockingTaskExecutor is not single-threaded anymore.