Fix memory leak which occurs when a client-side response stream is closed unexpectedly
Motivation:

A client-side response can be closed unexpectedly due to various
reasons, most notably due to a ResponseTimeoutException. When it
happens, an HTTP/1 connection has to be closed and an HTTP/2 stream has
to be reset with CANCEL error code. We don't do anything.

Modifications:

- Close the connection when a response of an HTTP/1 connection is closed
  unexpectedly.
- Reset the stream with CANCEL error code when a response of an HTTP/2
  connection is closed unexpectedly.

Result:

Memory leak is fixed.