Restore idle timeout handler to beginning of pipeline and use channel.close instead of ctx.close to make sure all handlers are run on idle timeout connection close. (#641)
With the current implementation, for HTTP/2 connections, IdleTimeoutHandler will not see any more channel reads after the initial handshake and will treat streaming requests as idle. Instead, we restore idle timeout handler to the beginning of the pipeline so it sees all channel reads, and have it close the channel, instead of the handler context, to ensure all outbound handlers are called, which will ensure logic like HTTP/2 goaway runs as described in #637

The regression test is quite slow and I'm worried about flakiness, it might not be worth it.