(WIP) Enhance annotated service path mapping: HTTP method and media type based mapping
Motivation:

Currently we cannot bind two or more service methods with the same path even if they have different HTTP methods. Also, we need to bind them depending on 'content-type' or 'accept' header of the request for some cases.
These modifications will affect only to annotated service for now, and then apply this feature to the other services later.

Modifications:

- Fix a bug that might get a wrong service instance from the cache when finding a service.
- Add PathMappingContext and use it PathMapping's argument instead of '(path, query)'.
- Introduce a simple content negotiation mechanism to find a service method by 'content-type' or 'accept' header.
- Add 'score' to PathMappingResult to find the best suitable service method.

Result:

- Fixes #579

To-Do:

- Apply caffeine cache to improve the efficiency of the service cache.
- Introduce some new service routing mechanism to find a service method more faster.