Fix EventLoop.terminationFuture() listener leak
Motivation:

HttpClientDelegate.pool() lazily creates a KeyedChannelPool which pools
TCP/IP connections. When a new pool is created, as a defensive measure,
pool() adds a listener to the terminationFuture() of EventLoop which
closes the all connections in the pool and removes itself from the map
of pools.

However, this can lead to a leak when a lot of HttpClientDelegates are
created because the listeners added to EventLoop.terminationFuture()
will linger until the EventLoop shuts down, which usually never happens
until the application shuts down.

It is also notable that it's not a good idea to close connections in
terminationFuture()'s listeners, because it's impossible to perform I/O
in a terminated EventLoop in Netty.

Modifications:

- Keep the number of HttpClientDelegates per HttpClientFactory to one
- Move the connection pool management from HttpClientDelegate to
  HttpClientFactory
- Use weak-key ConcurrentMap instead of ConcurrentHashMap so that the
  map of pools do not leak even when EventLoops are terminated and
  unreferenced externally
- Do not add a listener to terminationFuture()
- Double-check if the connection going back to a pool is active

Result:

Less leaks