Increase the chance of reusing idle client connections
Motivations:

Armeria currently keeps a client connection pool for each EventLoop. It
means, when a user sends a client request from N different EventLoops,
N new connections will be created. It is not a problem for light usage
but it can consume large amount of resources when the number of
EventLoops and the kinds of Endpoints get bigger.

Modifications:

- Add ClientFactory.acquireEventLoop(Endpoint) which enables a client
  implementation to choose the EventLoop which is more likely to have an
  idle Channel connected to the desired Endpoint.
- Add EventLoopScheduler which provides the implementation of
  acquireEventLoop(Endpoint)
- Use ClientFactory.acquireEventLoop(Endpoint) wherever possible
  - UserClient
  - ArmeriaChannel (gRPC)
- Miscellaneous:
  - Add ReleasableHolder
  - Fix IllegalStateException in Endpoint.equals(Object)
  - Reduce the memory footprint of PoolKey
  - Fix HttpTracingIntegrationTest so that it works even if the
    collected spans are not ordered

Result:

- The number of idle connections do not grow aggressively when the
  number of EventLoops and Endpoints are big.